apply Wojciech's suggestions
