+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoManager.kt
@@ -34,24 +34,8 @@ object VideoManager {
     return videoViews[id] ?: throw VideoViewNotFoundException(id)
   }
 
-  fun unregisterVideoView(videoView: VideoView, appContext: AppContext) {
+  fun unregisterVideoView(videoView: VideoView) {
     videoViews.remove(videoView.id)
-
-    // It is possible that the user configured their video to enter PiP mode automatically
-    // When they navigate away from the screen containing videos in that case, the PiP mode stays enabled by default
-    // Which would cause the entire activity to still go into PiP mode - it is safe to assume that the user wants
-    // the PiP behavior ONLY when they are watching a video.
-    // If there are no more VideoViews, we can assume we left the screen containing videos
-    // And therefore we should explicitly disable PiP mode just in case
-    if (videoViews.isEmpty()) {
-      val currentActivity = appContext.throwingActivity
-
-      if (Build.VERSION.SDK_INT >= 31 && isPictureInPictureSupported(currentActivity)) {
-        runWithPiPMisconfigurationSoftHandling {
-          currentActivity.setPictureInPictureParams(PictureInPictureParams.Builder().setAutoEnterEnabled(false).build())
-        }
-      }
-    }
   }
 
   fun registerVideoPlayer(videoPlayer: VideoPlayer) {
@@ -101,26 +85,5 @@ object VideoManager {
         videoView.videoPlayer?.player?.pause()
       }
     }
-  }
-
-  // We can't check if AndroidManifest.xml is configured properly, so we have to handle the exceptions ourselves to prevent crashes
-  fun runWithPiPMisconfigurationSoftHandling(shouldThrow: Boolean = false, ignore: Boolean = false, block: () -> Any?) {
-    try {
-      block()
-    } catch (e: IllegalStateException) {
-      if (ignore) {
-        return
-      }
-      Log.e("ExpoVideo", "Current activity does not support picture-in-picture. Make sure you have configured the `expo-video` config plugin correctly.")
-      if (shouldThrow) {
-        throw PictureInPictureConfigurationException()
-      }
-    }
-  }
-
-  fun isPictureInPictureSupported(currentActivity: Activity): Boolean {
-    return Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && currentActivity.packageManager.hasSystemFeature(
-      android.content.pm.PackageManager.FEATURE_PICTURE_IN_PICTURE
-    )
   }
 }
