+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoView.kt
@@ -7,6 +7,7 @@ import android.content.Intent
 import android.graphics.Canvas
 import android.graphics.Rect
 import android.os.Build
+import android.util.Log
 import android.util.Rational
 import android.view.View
 import android.view.ViewGroup
@@ -23,8 +24,7 @@ import com.facebook.yoga.YogaConstants
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.viewevent.EventDispatcher
 import expo.modules.kotlin.views.ExpoView
-import expo.modules.video.VideoManager.isPictureInPictureSupported
-import expo.modules.video.VideoManager.runWithPiPMisconfigurationSoftHandling
+import expo.modules.video.delegates.IgnoreSameSet
 import expo.modules.video.drawing.OutlineProvider
 import expo.modules.video.enums.ContentFit
 import expo.modules.video.player.VideoPlayer
@@ -80,15 +80,9 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
   private val borderDrawable
     get() = borderDrawableLazyHolder.value
 
-  var autoEnterPiP: Boolean = false
-    set(value) {
-      if (Build.VERSION.SDK_INT >= 31 && isPictureInPictureSupported(currentActivity) && field != value) {
-        runWithPiPMisconfigurationSoftHandling {
-          currentActivity.setPictureInPictureParams(PictureInPictureParams.Builder().setAutoEnterEnabled(value).build())
-        }
-      }
-      field = value
-    }
+  var autoEnterPiP: Boolean by IgnoreSameSet(false) {new, _ ->
+    applyAutoEnterPiP(new)
+  }
 
   var contentFit: ContentFit = ContentFit.CONTAIN
     set(value) {
@@ -323,6 +317,7 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
         .add(fragment, fragment.id)
         .commitAllowingStateLoss()
     }
+    applyAutoEnterPiP(autoEnterPiP);
   }
 
   override fun onDetachedFromWindow() {
@@ -334,6 +329,7 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
         .remove(fragment)
         .commitAllowingStateLoss()
     }
+    applyAutoEnterPiP(false);
   }
 
   @UnstableReactNativeAPI
@@ -405,6 +401,37 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
     if (shouldInvalided) {
       shouldInvalided = false
       invalidate()
+    }
+  }
+
+  // We can't check if AndroidManifest.xml is configured properly, so we have to handle the exceptions ourselves to prevent crashes
+  internal fun runWithPiPMisconfigurationSoftHandling(shouldThrow: Boolean = false, ignore: Boolean = false, block: () -> Any?) {
+    try {
+      block()
+    } catch (e: IllegalStateException) {
+      if (ignore) {
+        return
+      }
+      Log.e("ExpoVideo", "Current activity does not support picture-in-picture. Make sure you have configured the `expo-video` config plugin correctly.")
+      if (shouldThrow) {
+        throw PictureInPictureConfigurationException()
+      }
+    }
+  }
+
+  private fun applyAutoEnterPiP(autoEnterPiP: Boolean) {
+    if (Build.VERSION.SDK_INT >= 31 && isPictureInPictureSupported(currentActivity)) {
+      runWithPiPMisconfigurationSoftHandling {
+        currentActivity.setPictureInPictureParams(PictureInPictureParams.Builder().setAutoEnterEnabled(autoEnterPiP).build())
+      }
+    }
+  }
+
+  companion object {
+    fun isPictureInPictureSupported(currentActivity: Activity): Boolean {
+      return Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && currentActivity.packageManager.hasSystemFeature(
+        android.content.pm.PackageManager.FEATURE_PICTURE_IN_PICTURE
+      )
     }
   }
 }
